Refactor express app to own app.ts file
